id should be an integer

+++ web_module/branches/studio_or_changes_2007_07_11/src/main/com/topcoder/web/common/dao/ProjectDAO.java
@@ -1,16 +1,15 @@
 package com.topcoder.web.common.dao;
 
-import java.util.List;
-
 import com.topcoder.web.common.model.comp.Project;
 
+import java.util.List;
+
 /**
- * 
  * @author Cucu
  */
 public interface ProjectDAO {
 
-    public Project find(Long id);
-    
+    public Project find(Integer id);
+
     public List<Project> find(Integer statusId, Integer categoryId);
 }
